重新按行为拆分fg/js/popup.js
