--- src/bg/js/popup/fetchPushContent.js
@@ -1,6 +1,0 @@
-export default function fetchPushContent() {
-    chrome.storage.local.get(["AcpushList"], function (data) {
-      console.log(data);
-      $("#pop-push").append(data.AcpushList);
-    });
-  }
