--- src/bg/js/popup/renderMomentCircleHtml.js
@@ -1,6 +1,0 @@
-export default  function renderMomentCircleHtml() {
-    chrome.storage.local.get(["AcMomentCircle1"], function (data) {
-      $("#pop-push-momentcircle").append(data.AcMomentCircle1);
-    });
-  }
-
